fix(desktop): isolate smoke harness and harden regressions
Move smoke-only Electron bootstrap logic out of main so the desktop startup path stays readable while keeping the end-to-end checks in place. Align the smoke timeout layers and tighten runtime regression coverage around argon2 loading and desktop capabilities.
