+++ apps/web/tests/chat-stream-route.test.ts
@@ -116,6 +116,8 @@ describe('POST /api/w/[slug]/chat/stream', () => {
       teamManagement: true,
       connectors: true,
       kickstart: true,
+      autopilot: true,
+      slackIntegration: true,
     })
     mockGetSession.mockResolvedValue(session('alice'))
     mockFindCredentialsBySlug.mockResolvedValue({
@@ -198,6 +200,8 @@ describe('POST /api/w/[slug]/chat/stream', () => {
       teamManagement: false,
       connectors: true,
       kickstart: true,
+      autopilot: false,
+      slackIntegration: false,
     })
     mockGetSession.mockResolvedValue(session('local', 'ADMIN'))
     mockFindCredentialsBySlug.mockResolvedValue(null)
